feat: migrate erro response to v2 service
