--- src/services/edge-application-error-responses-services/index.js
@@ -1,4 +1,0 @@
-import { listErrorResponsesService } from './list-error-responses-service'
-import { editErrorResponsesService } from './edit-error-responses-service'
-
-export { listErrorResponsesService, editErrorResponsesService }
