--- src/services/edge-application-error-responses-services/make-edge-application-error-responses-base-url.js
@@ -1,4 +1,0 @@
-export const makeEdgeApplicationErrorResponsesBaseUrl = () => {
-  const version = 'v4'
-  return `${version}/edge_application/applications`
-}
