+++ src/views/EdgeApplications/TabsView.vue
@@ -31,7 +31,6 @@
     originsServices: { type: Object, required: true },
     clipboardWrite: { type: Function, required: true },
     deviceGroupsServices: { type: Object, required: true },
-    errorResponsesServices: { type: Object, required: true },
     rulesEngineServices: { type: Object, required: true },
     functionsServices: { type: Object, required: true },
     edgeFunctionsServices: { type: Object, required: true }
@@ -263,7 +262,6 @@
       condition: true,
       show: showTabs.errorResponses,
       props: () => ({
-        ...props.errorResponsesServices,
         edgeApplicationId: edgeApplicationId.value,
         listOriginsService: props.originsServices.listOriginsService
       })
